Revert "CHERRY-PICK: Fix "Extend host glibc workaround to clone3""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/688c644d-b6d4-4266-a4ed-486366f6f038). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic88f99aacac4f0197bf2e94a0dcf433a75bd4cb4