UPSTREAM: minijail: Introduce the ability to bypass multithreaded fork check
The multithreaded fork is onerous, as it can be considered (relatively)
safe under fork + exec or if the user is writing code in a fashion that's
async-safe. If the user is aware of the risks of fork, in Rust code
especially, then they can opt out.

See conversation:
https://groups.google.com/a/chromium.org/g/minijail/c/OE3NNw3QwRg

Test: m crosvm
Bug: 396763313

Change-Id: I737eb7986a014be4bd2a9c964e0345a859fe9af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minijail/+/5866409
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>