Use rust_bindgen module instead of pregenerated bindings
The libminijail-sys crate was using pregenerated bindings rather than
bindings generated at build from libminijail. This has probably been
fine since the crate lives alongside libminijail and so the bindings
likely match, however it only works because of a side effect in the
build system.

Soong by default sets OUT_DIR to `out`, which causes the
include to resolve correctly. However, OUT_DIR is configurable, and this
breaks if OUT_DIR is set to anything else. Further, this breaks RBE
build support as OUT_DIR is not set by default in RBE (Soong explicitly
sets it for Rust modules, but only when depending on generated source,
such as rust_bindgen bindings).

Bug: 407072297
Test: Remote RBE build passes.
Ignore-Upstream-First: These Android.bps seem to be developed in Android
Change-Id: I32ebc4a2d13dec7ac23b6b8dae745061216c8eed